Proof of concept for handling TILE_BOUNDING_BOX
